# fix: script parsing performance improvements