--- a/sonar-project.properties
+++ b/sonar-project.properties
@@ -1,4 +1,4 @@
-sonar.projectKey=bsv_go-sdk
+sonar.projectKey=bsv-blockchain_go-sdk
 sonar.organization=bsv-blockchain
 sonar.projectName=Go SDK
 sonar.projectDescription=BSV BLOCKCHAIN | Software Development Kit for Go