--- a/sonar-project.properties
+++ b/sonar-project.properties
@@ -1,5 +1,5 @@
-sonar.projectKey=bsv-blockchain_go-sdk
-sonar.organization=bsv-blockchain
+sonar.projectKey=bsv_go-sdk
+sonar.organization=bsv
 sonar.projectName=Go SDK
 sonar.projectDescription=BSV BLOCKCHAIN | Software Development Kit for Go
 sonar.projectVersion=1.0.0