## Summary

This PR fixes a performance regression in the opcode parser by implementing a two-pass approach that pre-allocates the exact capacity needed for parsed opcodes.

## Benchmark Comparison

This compares the performance of three different implementations of the opcode parser:

1. **Current** - Master branch as-is (no pre-allocation)
2. **Previous** - Master with line 138 modified to `parsedOps := make([]ParsedOpcode, 0, len(scr))`
3. **Optimized** - The fix/script-parsing-performance branch (two-pass approach with exact allocation)

## Benchmark Results

| Benchmark | Implementation | ns/op | B/op | allocs/op | Notes |
|-----------|---------------|-------|------|-----------|-------|
| **BenchmarkOpParseSmall** | | | | | *P2PKH script (25 bytes)* |
| | Current | 163.0 | 960 | 4 | |
| | Previous | 234.6 | 1,792 | 1 | |
| | **Optimized** | **83.09** | **320** | **1** | **49% faster than Current** |
| **BenchmarkOpParseLargeData** | | | | | *100KB data carrier script* |
| | Current | 46.66 | 192 | 2 | |
| | Previous | 88,971 | 6,406,160 | 1 | |
| | **Optimized** | **32.94** | **128** | **1** | **29% faster than Current** |
| **BenchmarkOpParseSuperLargeData** | | | | | *5MB data carrier script* |
| | Current | 40.05 | 192 | 2 | |
| | Previous | 1,494,738 | 320,004,125 | 2 | |
| | **Optimized** | **28.76** | **128** | **1** | **28% faster than Current** |
| **BenchmarkOpParseManyPushDatas** | | | | | *1000 push operations* |
| | Current | 19,782 | 145,988 | 11 | |
| | Previous | 140,293 | 6,529,045 | 1 | |
| | **Optimized** | **11,806** | **65,537** | **1** | **40% faster than Current** |
| **BenchmarkOpParseSTAS** | | | | | *Complex STAS script (1751 bytes)* |
| | Current | 29,133 | 244,295 | 12 | |
| | Previous | 15,742 | 114,691 | 1 | |
| | **Optimized** | **17,249** | **81,922** | **1** | **41% faster than Current** |